Generate poster ids

Given a poster googlesheet with the columns `poster_title` and `presenter_name` this function will create a new sheet that has a column with unique poster ids

generate_poster_ids(
  prefill_url,
  poster_key_file,
  dest_folder = "qr-codes",
  poster_id = NULL,
  pdf_compile = "all-qr-codes.pdf"
)

Arguments

prefill_url

Go to your form, click on the vertical "..." to see more options in the right corner. Click "get prefill link". Put in the poster id and poster name responses `{poster_id}` and `{poster_name}` respectively

poster_key_file

A link to a googlesheet or file path to a csv that contains at least two column names that are exactly `poster_title` and `presenter_name`

dest_folder

A character string of the folder where qr-codes should be saved. If the folder doesn't exist, one will be made with this name. Default is `qr-codes`.

poster_id

By default poster ids will be created. Alternatively you can use this argument to specify the column name of the existing IDs in the sheet you'd like to use.

pdf_compile

By default all QR codes created will be compiled into a PDF. This argument is a character that is the file path you'd like this PDF to be saved to.

Value

A PDF and folder of QR code PNGs, a data frame with the information, and googlesheets that has the links to those QR_codes and poster ids

Examples

if (FALSE) { # \dontrun{

# Declare the path to your data. Can be a googlesheet URL or a CSV file path
# You do not need to use paste to build your URL.
# We can't have lines longer than 100 char in our docs.
prefill_url <- paste0("https://docs.google.com/forms/d/e/",
"1FAIpQLSepdIqUoLA9fgPksF_x7r5-hvTbd8ZoDH2h0uUQuvVdvhujMA/viewform?usp=pp_url",
"&entry.38519462=%7Bposter_id%7D&entry.1154882998=%7Bposter_name%7D")

poster_googlesheet <-
"https://docs.google.com/spreadsheets/d/12aomFyT0zEHNmpyCQoGdDh16P-bRp4Pkt4PCCrU7gYY/edit#gid=0"

generate_poster_ids(prefill_url = prefill_url,
                   poster_key_file = poster_googlesheet)

poster_csv <- example_poster_data_file_path()

generate_poster_ids(prefill_url = prefill_url,
                   poster_key_file = poster_csv)
} # }